long * -> int *
git-svn-id: https://svn.r-project.org/R/branches/R-0-65-patches@5807 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/strsignif.c
@@ -94,7 +94,7 @@ void str_signif(char *x, int *n, char **type, int *width, int *digits,
 		}
 		if (strcmp("integer", *type) == 0)
 		  for (i=0; i < nn; i++)
-			sprintf(result[i], form, wid, (int)((long *)x)[i]);
+			sprintf(result[i], form, wid, ((int *)x)[i]);
 		else
 			error("`type' must be \"integer\" for  \"d\"-format\n");
 	}
